pyproject.toml
[project]
name = "m3terscan-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.123.2",
    "gql[all]>=4.0.0",
    "gunicorn>=23.0.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.38.0",
    "web3>=7.14.0",
]
